Move customer validation into github-event
Moving customer validation out of github-push
and into github-event depending on commit #271
