+++ github-event/Gopkg.toml
@@ -31,5 +31,5 @@
   unused-packages = true
 
 [[constraint]]
-  version = "0.5.9"
+  version = "0.8.1"
   name = "github.com/openfaas/openfaas-cloud"
